test: move OK test to common tests + general cleanup
