+++ k8s/tests/unit/test_charm_authorize_action.py
@@ -2,9 +2,8 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-import ops.testing as testing
 import vault.testing.authorize_action
-from vault.vault_client import AuditDeviceType
+from ops import testing
 
 from fixtures import MockBinding, VaultCharmFixtures
 
@@ -14,57 +13,8 @@ class TestCharmAuthorizeAction(VaultCharmFixtures, vault.testing.authorize_actio
         return [testing.Container(name="vault", can_connect=True)]
 
     def test_given_when_authorize_charm_then_charm_is_authorized(self):
-        mock_vault = self.mock_lib_vault
-        mock_vault.configure_mock(
-            **{
-                "authenticate.return_value": True,
-                "create_or_update_approle.return_value": "my-role-id",
-                "generate_role_secret_id.return_value": "my-secret-id",
-            },
-        )
         self.mock_get_binding.return_value = MockBinding(
             bind_address="1.2.3.4",
             ingress_address="1.2.3.4",
         )
-        container = testing.Container(
-            name="vault",
-            can_connect=True,
-        )
-        user_provided_secret = testing.Secret(
-            tracked_content={"token": "my token"},
-        )
-        peer_relation = testing.PeerRelation(
-            endpoint="vault-peers",
-        )
-        state_in = testing.State(
-            containers=[container],
-            leader=True,
-            secrets=[user_provided_secret],
-            relations=[peer_relation],
-        )
-        state_out = self.ctx.run(
-            self.ctx.on.action("authorize-charm", params={"secret-id": user_provided_secret.id}),
-            state=state_in,
-        )
-
-        mock_vault.enable_audit_device.assert_called_once_with(
-            device_type=AuditDeviceType.FILE, path="stdout"
-        )
-        mock_vault.enable_approle_auth_method.assert_called_once()
-        mock_vault.create_or_update_policy_from_file.assert_called_once_with(
-            name="charm-access",
-            path="src/templates/charm_policy.hcl",
-        )
-        mock_vault.create_or_update_approle.assert_called_once_with(
-            name="charm",
-            policies=["charm-access", "default"],
-            token_ttl="1h",
-            token_max_ttl="1h",
-        )
-        assert self.ctx.action_results == {
-            "result": "Charm authorized successfully. You may now remove the secret."
-        }
-        assert state_out.get_secret(label="vault-approle-auth-details").tracked_content == {
-            "role-id": "my-role-id",
-            "secret-id": "my-secret-id",
-        }
+        super().test_given_when_authorize_charm_then_charm_is_authorized()
